finalproject/requirements.txt
streamlit
llama-index
python-dotenv
pydantic
groq
jina
